pyproject.toml
[tool.poetry]
name = "histopathnet"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^7.2.0"
pandas = "^1.1.1"
matplotlib = "^3.3.1"
scikit-image = "^0.17.2"
tqdm = "^4.48.2"
joblib = "^0.16.0"
black = "^20.8b1"
imgaug = "^0.4.0"
opencv-python = "^4.4.0"
keras = "^2.4.3"
tensorflow-gpu = "^2.3.0"
jupyterlab = "^2.2.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
